# docs/api.yaml
openapi: 3.0.3
info:
  title: Spring Bank API
  version: 1.0.0
  description: |
    Banking simulation REST API for the Spring Bank demo application.

    ## Authentication
    Most endpoints require a valid **Bearer JWT** in the `Authorization` header.
    Obtain a token via `POST /api/auth/login` or `POST /api/auth/register`.

    Tokens expire after **15 minutes**. Use `POST /api/auth/refresh` with your
    `refreshToken` to obtain a new access token without re-authenticating.

    ## Error responses
    All errors follow the same envelope:
    ```json
    { "success": false, "error": "Human-readable message", "code": "MACHINE_CODE" }
    ```

    ## Monetary amounts
    All monetary values are expressed as **integer cents** (e.g. `$12.50` → `1250`)
    to avoid floating-point precision issues.
  contact:
    name: Spring Bank Demo
    email: support@springbank.demo
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api.springbank.demo
    description: Production demo
security:
  - bearerAuth: []
...
paths:
...
  /api/transactions/transfer:
    post:
      tags: [Transactions]
      summary: Internal account-to-account transfer
      description: |
        Moves funds between two Spring Bank accounts atomically.
        The source account must belong to the authenticated user and have
        sufficient available balance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fromAccountId, toAccountId, amountCents]
              properties:
                fromAccountId:
                  type: string
                  format: uuid
                toAccountId:
                  type: string
                  format: uuid
                amountCents:
                  type: integer
                  minimum: 1
                  example: 5000
                  description: "Amount in cents (e.g. 5000 = $50.00)"
                description:
                  type: string
                  maxLength: 255
                  example: "Rent payment"
      responses:
        "200":
          description: Transfer completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                error: "Insufficient funds."
                code: "INSUFFICIENT_FUNDS"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/transactions/external-transfer:
    post:
      tags: [Transactions]
      summary: Wire / ACH transfer to an external bank
      description: |
        Initiates an outbound transfer to an account at another financial institution.
        In the demo the transfer is simulated and immediately marked COMPLETED.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fromAccountId, amountCents, recipientName, routingNumber, externalAccountNumber]
              properties:
                fromAccountId:
                  type: string
                  format: uuid
                amountCents:
                  type: integer
                  minimum: 1
                  example: 100000
                recipientName:
                  type: string
                  example: "John Smith"
                routingNumber:
                  type: string
                  pattern: "^[0-9]{9}$"
                  example: "021000021"
                externalAccountNumber:
                  type: string
                  minLength: 4
                  maxLength: 17
                  example: "123456789"
                description:
                  type: string
                  maxLength: 255
      responses:
        "200":
          description: External transfer initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/transactions/bill-payment:
    post:
      tags: [Transactions]
      summary: Pay a registered biller
      description: |
        Debits a customer account and credits a registered biller account.
        Biller IDs are listed in the billers catalogue endpoint (not yet implemented).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fromAccountId, billerId, billerAccountNumber, amountCents]
              properties:
                fromAccountId:
                  type: string
                  format: uuid
                billerId:
                  type: string
                  example: "biller_electricity_001"
                billerAccountNumber:
                  type: string
                  example: "EL-98765432"
                amountCents:
                  type: integer
                  minimum: 1
                  example: 12000
                description:
                  type: string
                  maxLength: 255
      responses:
        "200":
          description: Bill payment processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/transactions/deposit:
    post:
      tags: [Transactions]
      summary: Record a cash or cheque deposit
      description: |
        Credits an account with the specified amount.  In production this would
        be triggered by a teller system; in the demo it can be called directly.
        Requires STAFF or ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [toAccountId, amountCents]
              properties:
                toAccountId:
                  type: string
                  format: uuid
                amountCents:
                  type: integer
                  minimum: 1
                  example: 50000
                description:
                  type: string
                  example: "Cash deposit at branch"
      responses:
        "200":
          description: Deposit recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/transactions/withdrawal:
    post:
      tags: [Transactions]
      summary: Record a cash withdrawal
      description: |
        Debits an account with the specified amount.
        Requires STAFF or ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fromAccountId, amountCents]
              properties:
                fromAccountId:
                  type: string
                  format: uuid
                amountCents:
                  type: integer
                  minimum: 1
                  example: 20000
                description:
                  type: string
                  example: "ATM withdrawal"
      responses:
        "200":
          description: Withdrawal recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Admin ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token obtained from `/api/auth/login`"
...
  schemas:
...
    TransactionType:
      type: string
      enum: [INTERNAL_TRANSFER, EXTERNAL_TRANSFER, BILL_PAYMENT, DEPOSIT, WITHDRAWAL]
    TransactionStatus:
      type: string
      enum: [PENDING, COMPLETED, FAILED, REVERSED]

    # ── Entities ─────────────────────────────────────────────────────────────
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        type:
          $ref: "#/components/schemas/TransactionType"
        status:
          $ref: "#/components/schemas/TransactionStatus"
        amountCents:
          type: integer
          minimum: 1
        currency:
          type: string
          example: "USD"
        fromAccountId:
          type: string
          format: uuid
          nullable: true
        toAccountId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        initiatedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required: [id, type, status, amountCents, currency, initiatedBy, createdAt, updatedAt]
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation failed."
        code:
          type: string
          example: "VALIDATION_ERROR"
      required: [success, error]
...
  responses:
    BadRequest:
      description: Validation error or malformed request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "amountCents must be a positive integer."
            code: "VALIDATION_ERROR"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "Authentication required."
            code: "AUTH_REQUIRED"
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "You do not have permission to perform this action."
            code: "FORBIDDEN"
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: "An unexpected error occurred."
            code: "INTERNAL_ERROR"